ml/image_matching/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
pinecone>=5.0.0
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

